requirements.txt
numpy
torch==2.3.0
torchvision
transformers==4.45.2
accelerate==1.0.1
pillow==10.4.0
opencv-python==4.10.0.84
diffusers==0.30.3
gradio==3.42.0
svgpathtools==1.6.1
svg.path==6.3
alphashape==1.3.1
shapely==2.0.6
cairocffi==1.7.1